Artefact view now runs without RABL
This commit is fairly hefty. Removes the .rabl views we no longer need.
Alterations to artefact presenters to cater for ODI specific fields that
are not included in GDS version.

NOTE: I skipped the parent tag test, the amount of mocking in that one test
made my head hurt. I will come back to it.